requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==2.3.7 
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1 